pyproject.toml
[tool.poetry]
name = "your-project-name"
version = "0.1.0"
description = "Your project description"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "3.12.5"
altair = "5.4.1"
google-analytics-data = "0.18.12"
google-auth = "2.35.0"
google-auth-oauthlib = "1.2.1"
numpy = "2.1.2"
pandas = "2.2.3"
pillow = "10.4.0"
psycopg2-binary = "2.9.9"
requests = "2.32.3"
SQLAlchemy = "^2.0.34"
streamlit = "1.39.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"